Got macOS booting

+++ build.sh
@@ -3,6 +3,11 @@
 IASL='./tools/iasl -vw 3073 -vi -vr -p'
 SRCACPI='src/ACPI'
 
+if [ ! -f src/smbios.txt ]; then
+    echo "Please generate your own serial number and put it in src/smbios.txt! (see src/smbios-sample.txt)"
+    exit 1
+fi
+
 rm -rf build && mkdir build
 
 # Copy OpenCore EFI folder
@@ -11,11 +16,14 @@ cp -R download/oc/OpenCorePkg/X64/EFI build
 OCFOLDER="build/EFI/OC"
 
 # Build ACPI
+$IASL $OCFOLDER/ACPI/SSDT-ALSD.aml $SRCACPI/SSDT-ALSD.dsl
 $IASL $OCFOLDER/ACPI/SSDT-AWAC.aml $SRCACPI/SSDT-AWAC.dsl
+$IASL $OCFOLDER/ACPI/SSDT-dGPU-Off.aml $SRCACPI/SSDT-dGPU-Off.dsl
 $IASL $OCFOLDER/ACPI/SSDT-DMAC.aml $SRCACPI/SSDT-DMAC.dsl
 $IASL $OCFOLDER/ACPI/SSDT-EC-USBX.aml $SRCACPI/SSDT-EC-USBX.dsl
 $IASL $OCFOLDER/ACPI/SSDT-GPRW.aml $SRCACPI/SSDT-GPRW.dsl
 $IASL $OCFOLDER/ACPI/SSDT-MEM2.aml $SRCACPI/SSDT-MEM2.dsl
+$IASL $OCFOLDER/ACPI/SSDT-NoHybGfx.aml $SRCACPI/SSDT-NoHybGfx.dsl
 $IASL $OCFOLDER/ACPI/SSDT-PLUG.aml $SRCACPI/SSDT-PLUG.dsl
 $IASL $OCFOLDER/ACPI/SSDT-PNLF-CFL.aml $SRCACPI/SSDT-PNLF-CFL.dsl
 $IASL $OCFOLDER/ACPI/SSDT-PPMC.aml $SRCACPI/SSDT-PPMC.dsl
@@ -28,17 +36,13 @@ cp -R download/drivers/* $OCFOLDER/Drivers/
 
 # Copy kexts
 cp -R download/kexts/* $OCFOLDER/Kexts/
-#cp -R src/Kexts/* $OCFOLDER/Kexts/
+cp -R src/Kexts/* $OCFOLDER/Kexts/
 
 # Copy OpenCore config
 cp src/config.plist $OCFOLDER/config.plist
 
 # Replace SMBIOS
-if [ -e src/smbios.txt ]; then
-    . src/smbios.txt
-else
-    . src/smbios-sample.txt
-fi
+. src/smbios.txt
 sed -i "" -e "s/MLB_PLACEHOLDER/$MLB/" \
           -e "s/Serial_PLACEHOLDER/$SystemSerialNumber/" \
           -e "s/SmUUID_PLACEHOLDER/$SystemUUID/" $OCFOLDER/config.plist
